Refactor: mudando nome do package entitys -> models
