+++ src/main/java/br/apae/ged/dto/DocumentRequestDTO.java
@@ -1,6 +1,6 @@
 package br.apae.ged.dto;
 
-import br.apae.ged.entitys.Document;
+import br.apae.ged.models.Document;
 import org.springframework.web.multipart.MultipartFile;
 
 public record DocumentRequestDTO(
