+++ src/main/java/br/apae/ged/models/enums/TipoDocumento.java
@@ -1,4 +1,4 @@
-package br.apae.ged.entitys.enums;
+package br.apae.ged.models.enums;
 
 public enum TipoDocumento {
